requirements.txt
aerich==0.7.2
asyncpg==0.28.0
fastapi==0.103.2
uvicorn==0.23.2
pydantic==2.4.2
pydantic-settings==2.0.3
tortoise-orm==0.20.0
psycopg2-binary==2.9.9
py-fastapi-logging==0.4.0